+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## 0.3.0
+### Added
+- Google Embed map.
+- Input for API key
+- Load API key from `${service}_api_key` setting.
+
+### Changed
+- Set default map dimensions to 400x300.
+- Auto adjust live map preview container size.
+- Changed minimum map dimension to 200px.
+- Only support `roadmap` and `satellite` map types.
+- Switched ownership to BrandExtract.
+
 ## 0.2.0 - 2016-02-01
 ### Added
 - Language pack with `en` for now.